Repository Pattern added using DAO layer
