+++ src/main/java/server/context/ServiceContext.java
@@ -60,6 +60,8 @@ public class ServiceContext {
 
     private static Scanner scanner;
 
+    // repository
+
     private ServiceContext(){
     }
