+++ src/main/java/server/service/impl/FileSystem/AppointServiceV1FileImpl.java
@@ -1,20 +1,25 @@
 package server.service.impl.FileSystem;
 
+import server.context.RepositoryContext;
 import server.domain.version1.AppointmentV1;
+import server.dao.impl.fileSystem.AppointmentV1RepoFileImpl;
 import server.service.version1.AppointmentServiceV1;
 
 import java.util.List;
 
 public class AppointServiceV1FileImpl implements AppointmentServiceV1 {
+    private AppointmentV1RepoFileImpl appointmentV1RepoFile;
+    public AppointServiceV1FileImpl(){
+        appointmentV1RepoFile = RepositoryContext.getAppointmentV1RepoFile();
+    }
     @Override
     public void saveAppointmentsToStorage(List<AppointmentV1> appointmentList) {
-        System.out.println("Filesystem is pending");
+        appointmentV1RepoFile.saveAppointmentsToStorage(appointmentList);
     }
 
     @Override
     public List<AppointmentV1> getAppointments() {
-        System.out.println("FileSystem is pending");
-        return null;
+        return appointmentV1RepoFile.getAppointments();
     }
 
     @Override
@@ -24,18 +29,16 @@ public class AppointServiceV1FileImpl implements AppointmentServiceV1 {
 
     @Override
     public void viewAppointmentsByPatientId(int patientId) {
-        System.out.println("Filesystem is pending");
-
+        appointmentV1RepoFile.viewAppointmentsByPatientId(patientId);
     }
 
     @Override
     public void viewAppointmentsByDoctorId(int doctorId) {
-        System.out.println("Filesystem is pending");
+        appointmentV1RepoFile.viewAppointmentsByDoctorId(doctorId);
     }
 
     @Override
     public boolean addAppointmentEntry(AppointmentV1 appointment) {
-        System.out.println("Filesystem is pending");
-        return false;
+        return appointmentV1RepoFile.addAppointmentEntry(appointment);
     }
 }
