+++ src/main/java/server/service/impl/FileSystem/UserServiceImpl.java
@@ -1,73 +1,43 @@
 package server.service.impl.FileSystem;
+import server.context.RepositoryContext;
 import server.domain.User;
 import server.service.UserService;
-import java.io.BufferedWriter;
-import java.io.FileWriter;
-import java.io.IOException;
-import java.nio.file.Files;
-import java.nio.file.Paths;
-import java.util.ArrayList;
+import server.dao.impl.fileSystem.UserRepoFileImpl;
+
 import java.util.List;
-import java.util.Optional;
-import java.util.stream.Collectors;
 
 public class UserServiceImpl implements UserService{
     public final static String userFilePath = "users.csv";
+    private final UserRepoFileImpl userRepoFile;
+
+    public UserServiceImpl(){
+        userRepoFile = RepositoryContext.getUserRepoFile();
+    }
 
     public List<User> getUsers() {
-        List<User> users = new ArrayList<>();
-        try {
-            List<String> lines = Files.readAllLines(Paths.get(UserServiceImpl.userFilePath));
-            for (String line : lines) {
-                String[] parts = line.split(",");
-                if (parts.length == 6) {
-                    int id = Integer.parseInt(parts[0]);
-                    String name = parts[1].strip();
-                    String username = parts[2].strip();
-                    String password = parts[3].strip();
-                    String roll = parts[4].strip();
-                    boolean accountLocked = Boolean.parseBoolean(parts[5]);
-
-                    User user = new User(id,name, roll, username, password,accountLocked);
-                    users.add(user);
-                }
-            }
-
-        } catch (IOException e) {
-            System.out.println("Error reading users from the file: " + e.getMessage());
-        }
-        return users;
+        return userRepoFile.getUsers();
     }
 
     @Override
     public boolean addUserEntry(User user) {
-        List<User> userList= getUsers();
-        userList.add(user);
-        addUsersListToStorage(userList);
-        return true;
+       return userRepoFile.addUserEntry(user);
     }
 
     // saves the users list to file. It overwrites the previous data with new data
     public void addUsersListToStorage(List<User> userList){
-        try (BufferedWriter writer = new BufferedWriter(new FileWriter(UserServiceImpl.userFilePath, false))) {
-            for (User itrUser : userList) {
-                writer.write(itrUser.getUserId() + "," + itrUser.getName() + "," + itrUser.getUsername() + "," + itrUser.getPassword() + "," + itrUser.getRole()
-                        + "," + itrUser.getAccountStatus());
-                writer.newLine();
-            }
-        } catch (IOException e) {
-            System.out.println("Error adding user to the file: " + e.getMessage());
-        }
+        userRepoFile.addUsersListToStorage(userList);
     }
 
     public User validateUserLogin(String username, String password, String userRole){
         // Logic to validate username and password
-        List<User> users = getUsers();
-        Optional<User> loginUser = users.stream()
-                .filter(user -> user.getUsername().equals(username) && user.getPassword().equals(password) )
-                .findFirst();
-        return loginUser.orElse(null);
-
+       User user = userRepoFile.getUserByUsername(username);
+       if(user==null){
+           return null;
+       }
+       else if(user.getPassword().equals(password) && user.getPassword().equals(password)){
+           return user;
+       }
+       return null;
     }
 
 
@@ -79,19 +49,10 @@ public class UserServiceImpl implements UserService{
         }
     }
     public List<User> getPatients() {
-        List<User> users = getUsers();
-        List<User> patients = users.stream()
-                .filter(user -> user.getRole().equals("patient"))
-                .collect(Collectors.toList());
-        return patients;
-
+        return userRepoFile.getPatients();
     }
     public List<User> getDoctors() {
-        List<User> users = getUsers();
-        List<User> doctors = users.stream()
-                .filter(user -> user.getRole().equals("doctor"))
-                .collect(Collectors.toList());
-        return doctors;
+        return userRepoFile.getDoctors();
     }
 
     public void viewPatients(){
@@ -110,7 +71,7 @@ public class UserServiceImpl implements UserService{
 
     @Override
     public boolean deleteUser(String username) {
-        return false;
+        return userRepoFile.deleteUser(username);
     }
