+++ src/main/java/server/service/impl/hibernate/UserServiceHibernateImpl.java
@@ -3,7 +3,9 @@ package server.service.impl.hibernate;
 import org.hibernate.Session;
 import org.hibernate.Transaction;
 import org.hibernate.exception.ConstraintViolationException;
+import server.context.RepositoryContext;
 import server.context.ServiceContext;
+import server.dao.impl.hibernate.UserRepoHibernateImpl;
 import server.domain.Slot;
 import server.domain.User;
 import server.domain.version1.AppointmentV1;
@@ -14,48 +16,20 @@ import java.util.List;
 import java.util.Optional;
 
 public class UserServiceHibernateImpl implements UserService {
+    private UserRepoHibernateImpl userRepoHibernate;
+
     public UserServiceHibernateImpl(){
+        userRepoHibernate = RepositoryContext.getUserRepoHibernate();
     }
-
-
     @Override
     public List<User> getUsers() {
-        Session session = ServiceContext.getSessionFactory().openSession();
-        List<User> userList = null;
-        // HQL
-        userList = session.createQuery("from User").list();
-        session.close();
-        return userList;
-
+        return userRepoHibernate.getUsers();
     }
 
 
     @Override
     public boolean addUserEntry(User user) {
-        Session session = ServiceContext.getSessionFactory().openSession();
-        Transaction transaction = null;
-        try{
-            // start the transaction
-            transaction =  session.beginTransaction();
-
-            // save student object
-            session.save(user);
-
-            //commit transaction
-            transaction.commit();
-        } catch (PersistenceException e){
-            if(e.getCause() instanceof ConstraintViolationException)
-                System.out.println("Username Already Exists:"+ e.getMessage());
-
-            if(transaction != null){
-                transaction.rollback();
-            }
-            session.close();
-            return false;
-
-        }
-        session.close();
-        return true;
+        return userRepoHibernate.addUserEntry(user);
     }
 
     @Override
@@ -65,39 +39,24 @@ public class UserServiceHibernateImpl implements UserService {
 
     @Override
     public User validateUserLogin(String username, String password, String userRole) {
-        Session session = ServiceContext.getSessionFactory().openSession();
-        User user = null;
-        //HQL Query
-        user = (User) session.createQuery("from User where username=:username AND password=:password " +
-                                    " AND role=:role")
-                .setParameter("username",username)
-                .setParameter("password",password)
-                .setParameter("role",userRole)
-                .uniqueResult();
-        session.close();
-        return user;
+        User user = userRepoHibernate.getUserByUsername(username);
+        if(user==null){
+            return null;
+        }
+        else if(user.getPassword().equals(password) && user.getPassword().equals(password)){
+            return user;
+        }
+        return null;
     }
 
     @Override
     public List<User> getPatients() {
-        Session session = ServiceContext.getSessionFactory().openSession();
-        List<User> patientList;
-        patientList = (List<User>) session.createQuery("from User where role=:role")
-                .setParameter("role","patient")
-                .list();
-        session.close();
-        return patientList;
+        return userRepoHibernate.getPatients();
     }
 
     @Override
     public List<User> getDoctors() {
-        Session session = ServiceContext.getSessionFactory().openSession();
-        List<User> doctorList;
-        doctorList = (List<User>) session.createQuery("from User where role=:role")
-                .setParameter("role","doctor")
-                .list();
-        session.close();
-        return doctorList;
+        return userRepoHibernate.getDoctors();
     }
 
     public void viewUsers() {
@@ -119,40 +78,6 @@ public class UserServiceHibernateImpl implements UserService {
 
     @Override
     public boolean deleteUser(String username) {
-        Session session = ServiceContext.getSessionFactory().openSession();
-        Transaction transaction = null;
-        try {
-            transaction = session.beginTransaction();
-            String hql = "FROM User as u WHERE u.username = :username";
-            User user = (User) session.createQuery(hql)
-                    .setParameter("username", username)
-                    .uniqueResult();
-            if(user==null) {
-                System.out.println("No User found against the provided username");
-                return false;
-            }
-
-
-            // Remove the association between AppointmentV1 and Slot entities
-            for (AppointmentV1 appointment : user.getAppointmentV1List()) {
-                Slot slot = appointment.getSlot();
-                slot.removeAppointmentV1(); // If we don`t remove the child appointment from parent slot, hibernate will through exception Exception in thread "main" javax.persistence.EntityNotFoundException: deleted object would be re-saved by cascade (remove deleted object from associations): [server.domain.version1.AppointmentV1#1]
-            }
-
-            // As we have already defined cascading to parent will automatically delete the children upon deletion, so no need
-            user.getAppointmentV1List().clear();
-            user.getSlots().clear();
-
-            session.delete(user);
-            transaction.commit();
-
-        } catch (Exception e){
-            System.out.println("Error: Unable to write data to Database using hibernate: " + e.getMessage());
-            e.printStackTrace();
-            if(transaction!=null){
-                transaction.rollback();
-            }
-        }
-        return false;
+        return userRepoHibernate.deleteUser(username);
     }
 }
